# Avoid type erasure where possible.