--- a/weft/src/extensions.rs
+++ b/weft/src/extensions.rs
@@ -17,19 +17,19 @@ impl<D: fmt::Display> Displayable for D {
 }
 
 impl<'a> WeftRenderable for &'a str {
-    fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+    fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
         target.text(self)
     }
 }
 
 impl WeftRenderable for String {
-    fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+    fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
         target.text(self)
     }
 }
 
 impl<'a, D: fmt::Display> WeftRenderable for Displayer<'a, D> {
-    fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+    fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
         target.text(&self.to_string())
     }
 }