--- a/weft/tests/api_for_renderable.rs
+++ b/weft/tests/api_for_renderable.rs
@@ -10,7 +10,7 @@ fn should_render_trivial_example() {
     struct TrivialExample;
     // This simulates what a template of the form `<p>Hello</p>` should compile to.
     impl WeftRenderable for TrivialExample {
-        fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+        fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
             target.start_element_attrs("p".into(), &[])?;
             target.text("Hello".into())?;
             target.end_element("p".into())?;
@@ -33,7 +33,7 @@ fn should_render_attrs() {
     struct TrivialExample;
     // This simulates what a template of the form `<p>Hello</p>` should compile to.
     impl WeftRenderable for TrivialExample {
-        fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+        fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
             target.start_element_attrs("p".into(), &[&AttrPair::new("class", "some-classes")])?;
             target.text("Hello".into())?;
             target.end_element("p".into())?;
@@ -57,7 +57,7 @@ fn render_supports_builtins() {
     struct TrivialExample;
     // This simulates what a template of the form `<p>Hello</p>` should compile to.
     impl WeftRenderable for TrivialExample {
-        fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+        fn render_to(&self, target: impl RenderTarget) -> Result<(), io::Error> {
             "Hello world!".render_to(target)?;
             Ok(())
         }
@@ -78,7 +78,7 @@ fn display_supports_renderable() {
     struct Displayable<D>(D);
     // This simulates what a template of the form `<p>Hello</p>` should compile to.
     impl<D: fmt::Display> WeftRenderable for Displayable<D> {
-        fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+        fn render_to(&self, target: impl RenderTarget) -> Result<(), io::Error> {
             use weft::prelude::*;
             self.0.display().render_to(target)?;
             Ok(())
@@ -95,7 +95,7 @@ fn display_supports_to_string() {
     struct Displayable<D>(D);
     // This simulates what a template of the form `<p>Hello</p>` should compile to.
     impl<D: fmt::Display> WeftRenderable for Displayable<D> {
-        fn render_to(&self, target: &mut dyn RenderTarget) -> Result<(), io::Error> {
+        fn render_to(&self, mut target: impl RenderTarget) -> Result<(), io::Error> {
             use weft::prelude::*;
             target.start_element_attrs(
                 "p".into(),