This should mean tht the compiler can do more monomorphisation, and hopefully better optimisations. This seems to result in a ~11%-24% speedup.

Before:

```
running 11 tests
test big_table_allocating_001x001   ... bench:       3,088 ns/iter (+/- 328) = 28 MB/s
test big_table_allocating_004x004   ... bench:      11,294 ns/iter (+/- 1,307) = 25 MB/s
test big_table_allocating_016x016   ... bench:     127,528 ns/iter (+/- 10,280) = 23 MB/s
test big_table_allocating_064x064   ... bench:   1,988,660 ns/iter (+/- 180,416) = 22 MB/s
test big_table_allocating_256x256   ... bench:  32,085,038 ns/iter (+/- 2,527,404) = 23 MB/s
test big_table_reuse_buffer_001x001 ... bench:       2,204 ns/iter (+/- 1,187) = 39 MB/s
test big_table_reuse_buffer_004x004 ... bench:       9,978 ns/iter (+/- 1,485) = 28 MB/s
test big_table_reuse_buffer_016x016 ... bench:     124,578 ns/iter (+/- 16,225) = 23 MB/s
test big_table_reuse_buffer_064x064 ... bench:   2,105,703 ns/iter (+/- 791,580) = 21 MB/s
test big_table_reuse_buffer_256x256 ... bench:  31,405,820 ns/iter (+/- 5,002,150) = 24 MB/s
test teams                          ... bench:       9,183 ns/iter (+/- 912) = 35 MB/s
```

After:
```
running 11 tests
test big_table_allocating_001x001   ... bench:       2,690 ns/iter (+/- 341) = 32 MB/s
test big_table_allocating_004x004   ... bench:       9,697 ns/iter (+/- 828) = 29 MB/s
test big_table_allocating_016x016   ... bench:     107,350 ns/iter (+/- 8,601) = 27 MB/s
test big_table_allocating_064x064   ... bench:   1,669,994 ns/iter (+/- 224,270) = 27 MB/s
test big_table_allocating_256x256   ... bench:  28,423,436 ns/iter (+/- 3,398,518) = 26 MB/s
test big_table_reuse_buffer_001x001 ... bench:       1,772 ns/iter (+/- 325) = 49 MB/s
test big_table_reuse_buffer_004x004 ... bench:       8,521 ns/iter (+/- 538) = 33 MB/s
test big_table_reuse_buffer_016x016 ... bench:     107,351 ns/iter (+/- 11,930) = 27 MB/s
test big_table_reuse_buffer_064x064 ... bench:   1,662,150 ns/iter (+/- 315,108) = 27 MB/s
test big_table_reuse_buffer_256x256 ... bench:  28,168,557 ns/iter (+/- 8,473,894) = 27 MB/s
test teams                          ... bench:       7,962 ns/iter (+/- 680) = 41 MB/s
```